microservices/api/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.2
flask-restplus==0.13.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.16.0
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
Werkzeug==0.16.1
zipp==3.1.0
